feat: move language services to package
