--- apps/helpwave/lib/config/language.dart
@@ -1,6 +1,0 @@
-/// Language configuration
-const List<Map<String, String>> languages = [
-  {"Name": "English", "Local": "US", "Language": "en"},
-  {"Name": "Deutsch", "Local": "DE", "Language": "de"}
-];
-const defaultLanguageIndex = 1;
