+++ apps/helpwave/lib/pages/settings_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:helpwave_localization/localization.dart';
+import 'package:helpwave_localization/localization_model.dart';
 import 'package:helpwave/styling/constants.dart';
 import 'package:provider/provider.dart';
 import 'package:helpwave/components/accept_dialog.dart';
@@ -7,7 +8,6 @@ import 'package:helpwave/pages/landing_page.dart';
 import 'package:helpwave/services/introduction_model.dart';
 import 'package:helpwave/pages/emergency_pass_page.dart';
 import 'package:helpwave/pages/language_selection_page.dart';
-import 'package:helpwave/services/language_model.dart';
 import 'package:helpwave/services/theme_model.dart';
 
 /// Page for displaying basic Settings
