+++ packages/helpwave_localization/lib/localization.dart
@@ -1,3 +1,4 @@
 export 'package:flutter_localizations/flutter_localizations.dart';
 export 'package:helpwave_localization/src/localization_impl.dart';
+export 'package:helpwave_localization/src/config/language.dart' show getSupportedLocals, getSupportedLocalsMap;
 export 'package:intl/intl.dart';
