+++ packages/helpwave_localization/lib/localization_model.dart
@@ -1,0 +1,1 @@
+export 'package:helpwave_localization/src/language_model.dart';
